Create main UI tabs skeleton and Reorganize Project Structure - Presentation Layer
