+++ app/build.gradle
@@ -70,16 +70,21 @@ dependencies {
     implementation "org.koin:koin-android-viewmodel:${koin}"
 
 
-
     // RxJava
     implementation "io.reactivex.rxjava2:rxjava:${rxJava2}"
     implementation "io.reactivex.rxjava2:rxandroid:${rxAndroid}"
 
 
-
-
     // LifeCycle
     implementation "androidx.lifecycle:lifecycle-extensions:${lifecycle}"
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'androidx.navigation:navigation-fragment:2.0.0'
+    implementation 'androidx.navigation:navigation-ui:2.0.0'
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.0.0'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     annotationProcessor "androidx.lifecycle:lifecycle-compiler:${lifecycle}"
     implementation "androidx.lifecycle:lifecycle-reactivestreams:${lifecycle}"
 
@@ -102,4 +107,3 @@ repositories {
     mavenCentral()
 
 }
-
