+++ app/src/main/java/com/levimoreira/cinemabuff/home/HomeFragment.kt
@@ -1,8 +1,11 @@
-package com.levimoreira.cinemabuff.movie.views
+package com.levimoreira.cinemabuff.home
 
 import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
 import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -10,20 +13,27 @@ import androidx.recyclerview.widget.RecyclerView
 import com.levimoreira.cinemabuff.R
 import com.levimoreira.cinemabuff.domain.entities.Movie
 import com.levimoreira.cinemabuff.domain.entities.NetworkState
-import com.levimoreira.cinemabuff.movie.viewmodels.MovieViewModel
-import kotlinx.android.synthetic.main.activity_main.*
+import kotlinx.android.synthetic.main.fragment_home.*
 import org.koin.android.viewmodel.ext.android.viewModel
 
-class MainActivity : AppCompatActivity() {
+class HomeFragment : Fragment() {
 
     val movieViewModel: MovieViewModel by viewModel()
 
     lateinit var moviesAdapter: MovieListAdapter
 
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+    override fun onCreateView(
+            inflater: LayoutInflater,
+            container: ViewGroup?,
+            savedInstanceState: Bundle?
+    ): View? {
 
+        return inflater.inflate(R.layout.fragment_home, container, false)
+    }
+
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
         setUpMovieList()
     }
 
@@ -32,10 +42,10 @@ class MainActivity : AppCompatActivity() {
         movieList.apply {
             adapter = moviesAdapter
             layoutManager =
-                    LinearLayoutManager(this@MainActivity, RecyclerView.VERTICAL, false)
+                    LinearLayoutManager(context, RecyclerView.VERTICAL, false)
         }
 
-        movieList.addItemDecoration(DividerItemDecoration(this, RecyclerView.VERTICAL))
+        movieList.addItemDecoration(DividerItemDecoration(context, RecyclerView.VERTICAL))
 
         moviesAdapter.onMovieClickListener = ::onMovieClick
 
@@ -54,6 +64,6 @@ class MainActivity : AppCompatActivity() {
     }
 
     private fun onMovieClick(movie: Movie) {
-        Toast.makeText(this, movie.title, Toast.LENGTH_SHORT).show()
+        Toast.makeText(context, movie.title, Toast.LENGTH_SHORT).show()
     }
 }
