+++ app/src/main/java/com/levimoreira/cinemabuff/infrastructure/application/CinemaBuffApp.kt
@@ -3,7 +3,7 @@ package com.levimoreira.cinemabuff.infrastructure.application
 import android.app.Application
 import com.cinemabuff.data.movieDataModule
 import com.cinemabuff.data.network.networkModule
-import com.levimoreira.cinemabuff.movie.movieModule
+import com.levimoreira.cinemabuff.infrastructure.movieModule
 import org.koin.android.ext.android.startKoin
 
 class CinemaBuffApp : Application() {
