+++ app/src/main/java/com/levimoreira/cinemabuff/navigation/BlankFragment.kt
@@ -1,0 +1,24 @@
+package com.levimoreira.cinemabuff.navigation
+
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+import com.levimoreira.cinemabuff.R
+
+/**
+ * A simple [Fragment] subclass.
+ */
+class BlankFragment : Fragment() {
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
+                              savedInstanceState: Bundle?): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_blank, container, false)
+    }
+
+
+}
