+++ app/src/main/res/drawable/ic_favorite_black_24dp.xml
@@ -1,0 +1,4 @@
+<vector android:height="24dp" android:viewportHeight="192"
+    android:viewportWidth="192" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#FF000000" android:pathData="m192,96c0,-53.02 -42.98,-96 -96,-96s-96,42.98 -96,96 42.98,96 96,96h96v-16h-43c26.84,-17.77 42.988,-47.809 43,-80zM143.023,160.723c-10.723,7.789 -25.734,5.414 -33.523,-5.313 -7.789,-10.723 -5.414,-25.73 5.309,-33.523 10.727,-7.789 25.734,-5.414 33.523,5.313 7.793,10.723 5.414,25.73 -5.309,33.523zM141.844,55.871c12.605,-4.098 26.145,2.801 30.242,15.406 4.094,12.609 -2.805,26.148 -15.41,30.242 -12.605,4.098 -26.145,-2.801 -30.242,-15.406 -4.094,-12.605 2.805,-26.148 15.41,-30.242zM96,16c13.254,0 24,10.746 24,24s-10.746,24 -24,24 -24,-10.746 -24,-24 10.746,-24 24,-24zM19.914,71.277c4.098,-12.605 17.637,-19.504 30.242,-15.41 12.605,4.098 19.504,17.637 15.41,30.242 -4.098,12.605 -17.637,19.504 -30.242,15.41s-19.504,-17.633 -15.41,-30.242zM82.5,155.41c-7.789,10.727 -22.801,13.102 -33.523,5.313 -10.723,-7.793 -13.102,-22.801 -5.309,-33.523 7.789,-10.723 22.801,-13.102 33.523,-5.309 10.723,7.789 13.098,22.797 5.309,33.52zM88,96c0,-4.418 3.582,-8 8,-8s8,3.582 8,8 -3.582,8 -8,8 -8,-3.582 -8,-8zM88,96"/>
+</vector>
