+++ app/src/main/res/drawable/ic_home_black_24dp.xml
@@ -1,9 +1,18 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
     android:width="24dp"
     android:height="24dp"
-    android:viewportHeight="24.0"
-    android:viewportWidth="24.0">
+    android:viewportWidth="23.382"
+    android:viewportHeight="23.382">
     <path
         android:fillColor="#FF000000"
-        android:pathData="M10,20v-6h4v6h5v-8h3L12,3 2,12h3v8z" />
+        android:pathData="M14.58,15.181c0.438,-0.732 1.082,-1.107 1.936,-1.117c0.854,-0.01 1.285,-0.445 1.299,-1.301c0.012,-0.852 0.383,-1.496 1.115,-1.932c0.734,-0.438 0.893,-1.027 0.475,-1.773c-0.416,-0.744 -0.416,-1.488 0,-2.234c0.418,-0.744 0.26,-1.332 -0.475,-1.77c-0.732,-0.439 -1.104,-1.082 -1.115,-1.938c-0.014,-0.852 -0.445,-1.287 -1.299,-1.297c-0.854,-0.012 -1.498,-0.385 -1.936,-1.117c-0.438,-0.734 -1.027,-0.893 -1.771,-0.475c-0.744,0.416 -1.49,0.416 -2.234,0C9.83,-0.19 9.241,-0.032 8.803,0.702C8.366,1.435 7.721,1.808 6.868,1.819c-0.852,0.01 -1.285,0.445 -1.297,1.297C5.557,3.972 5.186,4.614 4.454,5.054C3.72,5.492 3.559,6.079 3.979,6.824c0.418,0.746 0.418,1.49 0,2.234c-0.42,0.746 -0.26,1.336 0.475,1.773c0.732,0.436 1.104,1.08 1.117,1.932c0.012,0.855 0.445,1.291 1.297,1.301c0.854,0.01 1.498,0.385 1.936,1.117c0.438,0.734 1.027,0.893 1.771,0.473c0.744,-0.412 1.49,-0.412 2.234,0C13.553,16.073 14.143,15.915 14.58,15.181zM11.694,12.614c-2.582,0 -4.674,-2.092 -4.674,-4.674c0,-2.58 2.092,-4.672 4.674,-4.672c2.58,0 4.672,2.092 4.672,4.672C16.366,10.522 14.274,12.614 11.694,12.614z" />
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M6.793,14.749c-0.898,0 -1.018,-0.418 -1.018,-0.418l-3.507,6.893l2.812,-0.316l1.555,2.34c0,0 3.24,-6.76 3.24,-6.715C8.196,16.491 8.864,14.794 6.793,14.749z" />
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M17.563,14.601c-2.562,0.268 -2.041,0.904 -2.627,1.398c-0.674,0.719 -1.516,0.578 -1.516,0.578l3.955,6.805l0.973,-2.52l2.766,0.361L17.563,14.601z" />
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M12.67,6.909l-0.978,-1.98l-0.979,1.98l-2.189,0.32l1.582,1.543l-0.373,2.182l1.959,-1.029l1.959,1.029l-0.373,-2.182l1.582,-1.543z" />
 </vector>
