+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/mobile_navigation"
+    app:startDestination="@+id/navigation_home">
+
+    <fragment
+        android:id="@+id/navigation_home"
+        android:name="com.levimoreira.cinemabuff.home.HomeFragment"
+        android:label="@string/title_home"
+        tools:layout="@layout/fragment_home" />
+    <fragment
+        android:id="@+id/navigation_suggested"
+        android:name="com.levimoreira.cinemabuff.navigation.BlankFragment"
+        android:label="@string/title_suggested_movie"
+        tools:layout="@layout/fragment_blank" />
+    <fragment
+        android:id="@+id/navigation_favorites"
+        android:name="com.levimoreira.cinemabuff.navigation.BlankFragment"
+        android:label="@string/title_favorites"
+        tools:layout="@layout/fragment_blank" />
+
+</navigation>
