+++ app/src/main/res/values/styles.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <style name="Theme.CinemaBuff" parent="Theme.AppCompat.Light">
-        <item name="colorPrimary">#ffffff</item>
-        <item name="colorPrimaryDark">#e0e0e0</item>
-        <item name="colorAccent">#607d8b</item>
+    <style name="Theme.CinemaBuff" parent="Theme.AppCompat.NoActionBar">
+        <item name="colorPrimary">#d50000</item>
+        <item name="colorPrimaryDark">#9b0000</item>
+        <item name="colorAccent">#262626</item>
     </style>
 
     <style name="Movie.Title" parent="TextAppearance.AppCompat">
